Move exception.hpp out of include/deepstream
