+++ src/exception.cpp
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include <deepstream/exception.hpp>
+#include "exception.hpp"
 
 namespace deepstream {
